add loger to TASK

+++ implementation/tasks/delete_dir_task.py
@@ -1,8 +1,13 @@
 import os
 from core.task_base import TaskBase
+from scheduler_config import get_logger
+
+logger = get_logger()
 
 
 class DeleteDirTask(TaskBase):
+    """Класс для удаления папки."""
+
     def __init__(self, param=None):
         super().__init__(param)
 
@@ -10,6 +15,9 @@ class DeleteDirTask(TaskBase):
     def name(self):
         return 'delete dir'
 
-    def execute(self):
-        os.rmdir(self.param)
-        print(f'Delete dir  {self.param}')
+    def execute(self) -> None:
+        if os.path.is_dir(self.param):
+            os.rmdir(self.param)
+            print(f'Delete dir  {self.param}')
+            logger.info('Deleted %s.', self.param)
+        logger.info('Finished deleting dir.')
